refactor(quality): split oversized production modules
