+++ docs/iterations/06-production-hardening/64-production-parity-hardening-plan.md
@@ -207,6 +207,19 @@ Evidence:
 - full allowed `verify` surface, excluding prohibited real-project agent or
   runtime smoke flows.
 
+Maintainability implementation evidence:
+
+- all ten baseline production files now stay below the 500-line quality budget;
+- provider-specific query, pagination/window, metrics and serialization logic
+  lives in bounded companion modules while existing public imports remain
+  stable;
+- the Prisma transaction guard now accepts either quote style and still
+  requires `Serializable` together with the shared P2034 retry boundary;
+- the persistence-readiness inventory exactly matches current Monitoring and
+  Ingestion Worker in-memory adapters;
+- TypeScript, 144 focused Jest tests, changed-file ESLint, code-quality,
+  persistence-readiness and source-line-cap gates pass.
+
 ## Likely Owned Edit Set
 
 - `libs/monitoring/features/request-scan/**`
