+++ libs/ingestion/adapters/source/rss/rss-source.provider.ts
@@ -25,10 +25,14 @@ import type {
   RssClientPort,
   RssFeedItem,
   RssReadFeedOptions,
-  RssReadFeedResult,
 } from "./rss-client.port";
-
-
+import {
+  feedUrlsForTargetWindow,
+  filterItemsForWindow,
+  filterSourcedItemsByRelativeAge,
+  readTargetPublishedWindow,
+  type SourcedRssFeedItem,
+} from "./rss-source-window";
 type RssFeedReadResult =
   | {
       readonly ok: true;
@@ -39,17 +43,6 @@ type RssFeedReadResult =
       readonly feedUrl: string;
       readonly error: unknown;
     };
-
-type SourcedRssFeedItem = {
-  readonly feedUrl: string;
-  readonly item: RssFeedItem;
-  readonly index: number;
-};
-
-type TargetPublishedWindow = {
-  readonly startInclusive: Date;
-  readonly endExclusive: Date;
-};
 
 const capabilityProfile: SourceCapabilityProfile = {
   providerKey: "rss",
@@ -329,98 +322,6 @@ const searchQueryFromFeedUrl = (feedUrl: string): string | undefined => {
   }
 };
 
-const readTargetPublishedWindow = (
-  config: SourceProviderScanContext["config"],
-): TargetPublishedWindow | undefined => {
-  const raw = config?.targetPublishedWindow;
-  if (raw === null || typeof raw !== "object" || Array.isArray(raw)) {
-    return undefined;
-  }
-
-  const record = raw as Record<string, unknown>;
-  const start = readDate(record.startInclusive);
-  const end = readDate(record.endExclusive);
-
-  return start === undefined ||
-    end === undefined ||
-    start.getTime() >= end.getTime()
-    ? undefined
-    : { startInclusive: start, endExclusive: end };
-};
-
-const readDate = (value: unknown): Date | undefined => {
-  if (typeof value !== "string") {
-    return undefined;
-  }
-  const date = new Date(value);
-
-  return Number.isNaN(date.getTime()) ? undefined : date;
-};
-
-const feedUrlsForTargetWindow = (
-  feedUrls: readonly string[],
-  targetWindow: TargetPublishedWindow | undefined,
-): readonly string[] =>
-  targetWindow === undefined
-    ? feedUrls
-    : feedUrls.flatMap((feedUrl) =>
-        googleNewsHistoricalFeedUrls(feedUrl, targetWindow) ?? [feedUrl],
-      );
-
-const googleNewsHistoricalFeedUrls = (
-  feedUrl: string,
-  targetWindow: TargetPublishedWindow,
-): readonly string[] | undefined => {
-  try {
-    const parsed = new URL(feedUrl);
-    if (
-      parsed.hostname !== "news.google.com" ||
-      parsed.pathname !== "/rss/search"
-    ) {
-      return undefined;
-    }
-    const query = parsed.searchParams.get("q")?.trim();
-    if (query === undefined || query.length === 0) {
-      return undefined;
-    }
-    const queryTerms = historicalGoogleNewsQueryTerms(query);
-
-    return queryTerms.map((term) => {
-      const historicalUrl = new URL(parsed.toString());
-      historicalUrl.searchParams.set(
-        "q",
-        `${term} after:${dateToken(targetWindow.startInclusive)} before:${dateToken(
-          targetWindow.endExclusive,
-        )}`,
-      );
-
-      return historicalUrl.toString();
-    });
-  } catch {
-    return undefined;
-  }
-};
-
-const maxHistoricalGoogleNewsFeeds = 12;
-
-const historicalGoogleNewsQueryTerms = (query: string): readonly string[] => {
-  const normalized = queryWithoutRelativeWindow(query);
-  const terms = normalized
-    .split(/\s+OR\s+/iu)
-    .map((term) => term.trim())
-    .filter((term) => term.length > 0);
-
-  return (terms.length === 0 ? [normalized] : terms).slice(
-    0,
-    maxHistoricalGoogleNewsFeeds,
-  );
-};
-
-const queryWithoutRelativeWindow = (query: string): string =>
-  query.replace(/\bwhen:\d+[dhm]\b/giu, "").replace(/\s+/gu, " ").trim();
-
-const dateToken = (date: Date): string => date.toISOString().slice(0, 10);
-
 const rssWarnings = (items: readonly RssFeedItem[]): readonly string[] => [
   ...(items.some(
     (item) =>
@@ -445,40 +346,6 @@ const rssWarnings = (items: readonly RssFeedItem[]): readonly string[] => [
     : []),
 ];
 
-const filterItemsByRelativeAge = (
-  items: readonly RssFeedItem[],
-  maxItemAgeHours: number | undefined,
-): readonly RssFeedItem[] => {
-  if (maxItemAgeHours === undefined) {
-    return items;
-  }
-
-  const newestPublishedAt = items.reduce<number | undefined>((newest, item) => {
-    const publishedTime = item.publishedAt?.getTime();
-
-    if (publishedTime === undefined) {
-      return newest;
-    }
-
-    return newest === undefined
-      ? publishedTime
-      : Math.max(newest, publishedTime);
-  }, undefined);
-
-  if (newestPublishedAt === undefined) {
-    return items;
-  }
-
-  const minimumPublishedAt =
-    newestPublishedAt - maxItemAgeHours * 60 * 60 * 1000;
-
-  return items.filter(
-    (item) =>
-      item.publishedAt === undefined ||
-      item.publishedAt.getTime() >= minimumPublishedAt,
-  );
-};
-
 const rssRecencyWarnings = (
   originalItems: readonly RssFeedItem[],
   filteredItems: readonly RssFeedItem[],
@@ -489,38 +356,6 @@ const rssRecencyWarnings = (
         `Some RSS items were older than maxItemAgeHours=${maxItemAgeHours}; they were skipped.`,
       ]
     : [];
-
-const filterItemsForWindow = (
-  items: readonly RssFeedItem[],
-  maxItemAgeHours: number | undefined,
-  targetWindow: TargetPublishedWindow | undefined,
-): readonly RssFeedItem[] =>
-  targetWindow === undefined
-    ? filterItemsByRelativeAge(items, maxItemAgeHours)
-    : items.filter((item) => isInsideTargetWindow(item, targetWindow));
-
-const isInsideTargetWindow = (
-  item: RssFeedItem,
-  targetWindow: TargetPublishedWindow,
-): boolean =>
-  item.publishedAt === undefined ||
-  (item.publishedAt >= targetWindow.startInclusive &&
-    item.publishedAt < targetWindow.endExclusive);
-
-const filterSourcedItemsByRelativeAge = (
-  entries: readonly SourcedRssFeedItem[],
-  maxItemAgeHours: number | undefined,
-  targetWindow: TargetPublishedWindow | undefined,
-): readonly SourcedRssFeedItem[] => {
-  const filteredItems = filterItemsForWindow(
-    entries.map(({ item }) => item),
-    maxItemAgeHours,
-    targetWindow,
-  );
-  const filteredSet = new Set(filteredItems);
-
-  return entries.filter(({ item }) => filteredSet.has(item));
-};
 
 const hasReadableContent = (item: RssFeedItem): boolean =>
   (item.title ?? "").trim().length + (item.content ?? "").trim().length > 0;
