+++ libs/ingestion/adapters/source/source-query-plan-runtime-compiler.ts
@@ -6,6 +6,13 @@ import type {
   SourceQueryPlannerIntent,
 } from "../../domain";
 import type { SourceQuery, SourceRuntimeConfig } from "../../ports";
+import { compileRedditPlannedQuery } from "./source-query-plan-reddit-runtime";
+import {
+  executableLanes,
+  fallbackCompilation,
+  maxItemsForLane,
+  readStringArrayFromValues,
+} from "./source-query-plan-runtime-support";
 import {
   compactRuntimeConfig,
   compactUnique,
@@ -128,53 +135,6 @@ export const sourceQueryPlannerIntentFromConfig = (params: {
   };
 };
 
-const compileRedditPlannedQuery = (params: {
-  readonly originalSourceQuery: SourceQuery;
-  readonly runtimeConfig?: SourceRuntimeConfig;
-  readonly plan: SourceQueryPlan;
-}): SourceQueryPlannerRuntimeCompilation => {
-  const sourceLanes = executableLanes(params.plan, "reddit");
-  const scanLanes = sourceLanes.filter(
-    (lane) => lane.operation !== "enrichment",
-  );
-  const enrichmentLanes = sourceLanes.filter(
-    (lane) => lane.operation === "enrichment",
-  );
-  const baselineScanPasses = redditBaselineScanPassesFromConfig(
-    params.runtimeConfig,
-  );
-  const allowedSubreddits = redditAllowedSubredditsForPlan(
-    scanLanes,
-    params.runtimeConfig,
-  );
-  const scanPasses = mergeRedditScanPasses(
-    scanLanes.flatMap((lane) => redditScanPassForLane(lane, allowedSubreddits)),
-    baselineScanPasses,
-  );
-  const primaryLane = scanLanes[0];
-
-  if (scanPasses.length === 0 || primaryLane === undefined) {
-    return fallbackCompilation(
-      params.originalSourceQuery,
-      "source_query_planner.no_reddit_scan_passes",
-    );
-  }
-
-  return {
-    sourceQuery: {
-      mode: "search",
-      query: primaryLane.query,
-      parameters: compactRuntimeConfig({
-        maxItems: redditCompilationMaxItems(scanLanes, params.runtimeConfig),
-        scanPasses,
-        ...redditCommentExpansionParameters(enrichmentLanes),
-      }),
-    },
-    warnings: skippedLaneWarnings(sourceLanes, scanPasses.length),
-    applied: true,
-  };
-};
-
 const compileXTwitterPlannedQuery = (params: {
   readonly originalSourceQuery: SourceQuery;
   readonly runtimeConfig?: SourceRuntimeConfig;
@@ -231,223 +191,6 @@ const compileXTwitterPlannedQuery = (params: {
   };
 };
 
-const executableLanes = (
-  plan: SourceQueryPlan,
-  providerKey: string,
-): readonly SourceQueryPlanLane[] =>
-  plan.lanes.filter((lane) => lane.sourceKey === providerKey);
-
-const redditScanPassForLane = (
-  lane: SourceQueryPlanLane,
-  fallbackAllowedSubreddits?: readonly string[],
-): readonly SourceRuntimeConfig[] => {
-  if (lane.operation === "listing") {
-    const listing = redditListingFromQuery(lane.query);
-
-    return listing === undefined
-      ? []
-      : [
-          compactRuntimeConfig({
-            mode: "listing",
-            subreddit: listing.subreddit,
-            listing: listing.listing,
-            maxItems: maxItemsForLane(lane),
-            topTime: readStringParameter(lane, "topTime"),
-          }),
-        ];
-  }
-
-  if (lane.operation !== "search") {
-    return [];
-  }
-
-  return [
-    compactRuntimeConfig({
-      mode: "search",
-      query: lane.query,
-      maxItems: maxItemsForLane(lane),
-      searchSort:
-        readStringParameter(lane, "searchSort") ??
-        (lane.kind === "general" ? "new" : undefined),
-      searchTime: readStringParameter(lane, "searchTime"),
-      allowedSubreddits:
-        readStringArrayParameter(lane, "allowedSubreddits") ??
-        fallbackAllowedSubreddits,
-    }),
-  ];
-};
-
-const redditBaselineScanPassesFromConfig = (
-  runtimeConfig: SourceRuntimeConfig | undefined,
-): readonly SourceRuntimeConfig[] =>
-  readArray(runtimeConfig?.scanPasses).flatMap((pass) => {
-    const record = readRecord(pass);
-    const mode = readString(record?.mode) ?? "listing";
-
-    if (mode === "listing") {
-      const subreddit = readString(record?.subreddit ?? record?.query);
-      if (subreddit === undefined) {
-        return [];
-      }
-
-      return [
-        compactRuntimeConfig({
-          mode: "listing",
-          subreddit,
-          listing: readString(record?.listing) ?? "top",
-          topTime: readString(record?.topTime),
-          maxItems: readFiniteNumber(record?.maxItems),
-          minScore: readFiniteNumber(record?.minScore),
-          includeComments: readBoolean(record?.includeComments),
-          maxCommentsPerPost: readFiniteNumber(record?.maxCommentsPerPost),
-        }),
-      ];
-    }
-
-    if (mode !== "search") {
-      return [];
-    }
-
-    const query = readString(record?.query);
-    if (query === undefined) {
-      return [];
-    }
-
-    return [
-      compactRuntimeConfig({
-        mode: "search",
-        query,
-        searchSort: readString(record?.searchSort),
-        searchTime: readString(record?.searchTime),
-        maxItems: readFiniteNumber(record?.maxItems),
-        minScore: readFiniteNumber(record?.minScore),
-        includeComments: readBoolean(record?.includeComments),
-        maxCommentsPerPost: readFiniteNumber(record?.maxCommentsPerPost),
-        allowedSubreddits: readStringArrayFromValues(
-          record?.allowedSubreddits,
-          record?.subreddits,
-        ),
-      }),
-    ];
-  });
-
-const mergeRedditScanPasses = (
-  planned: readonly SourceRuntimeConfig[],
-  baseline: readonly SourceRuntimeConfig[],
-): readonly SourceRuntimeConfig[] =>
-  compactUniqueBy([...planned, ...baseline], redditScanPassKey);
-
-const redditScanPassKey = (pass: SourceRuntimeConfig): string => {
-  const mode = readString(pass.mode) ?? "listing";
-  if (mode === "search") {
-    return [
-      "search",
-      readString(pass.query) ?? "",
-      readString(pass.searchSort) ?? "",
-      readString(pass.searchTime) ?? "",
-      readStringArrayFromValues(pass.allowedSubreddits, pass.subreddits)
-        .map((subreddit) => subreddit.toLowerCase())
-        .join(","),
-    ].join(":");
-  }
-
-  return [
-    "listing",
-    (readString(pass.subreddit ?? pass.query) ?? "").toLowerCase(),
-    readString(pass.listing) ?? "top",
-    readString(pass.topTime) ?? "",
-  ].join(":");
-};
-
-const redditCompilationMaxItems = (
-  lanes: readonly SourceQueryPlanLane[],
-  runtimeConfig: SourceRuntimeConfig | undefined,
-): number => {
-  const configuredMaxItems = readOptionalPositiveInteger(
-    runtimeConfig?.maxItems,
-  );
-
-  return Math.min(100, Math.max(totalMaxItems(lanes), configuredMaxItems ?? 1));
-};
-
-const readFiniteNumber = (value: unknown): number | undefined =>
-  typeof value === "number" && Number.isFinite(value) ? value : undefined;
-
-const redditListingFromQuery = (
-  query: string,
-): { readonly subreddit: string; readonly listing: string } | undefined => {
-  const [subreddit, listing] = query.split(":");
-
-  if (
-    subreddit === undefined ||
-    subreddit.trim().length === 0 ||
-    listing === undefined ||
-    listing.trim().length === 0
-  ) {
-    return undefined;
-  }
-
-  return {
-    subreddit: subreddit.trim(),
-    listing: listing.trim(),
-  };
-};
-
-const redditAllowedSubredditsForPlan = (
-  lanes: readonly SourceQueryPlanLane[],
-  runtimeConfig: SourceRuntimeConfig | undefined,
-): readonly string[] | undefined => {
-  const subreddits = [
-    ...redditAllowedSubredditsFromConfig(runtimeConfig),
-    ...lanes.flatMap((lane) => {
-      if (lane.operation !== "listing") {
-        return [];
-      }
-
-      const listing = redditListingFromQuery(lane.query);
-
-      return listing === undefined ? [] : [listing.subreddit];
-    }),
-  ];
-
-  return subreddits.length === 0 ? undefined : compactUnique(subreddits);
-};
-
-const redditAllowedSubredditsFromConfig = (
-  runtimeConfig: SourceRuntimeConfig | undefined,
-): readonly string[] => [
-  ...readStringArrayFromValues(runtimeConfig?.allowedSubreddits),
-  ...readArray(runtimeConfig?.scanPasses).flatMap((pass) => {
-    const record = readRecord(pass);
-    const mode = readString(record?.mode) ?? "listing";
-    const subreddit = readString(record?.subreddit ?? record?.query);
-
-    if (mode === "search") {
-      return readStringArrayFromValues(record?.allowedSubreddits);
-    }
-
-    return mode === "listing" && subreddit !== undefined ? [subreddit] : [];
-  }),
-];
-
-const redditCommentExpansionParameters = (
-  lanes: readonly SourceQueryPlanLane[],
-): SourceRuntimeConfig => {
-  const lane = lanes[0];
-
-  if (lane === undefined) {
-    return {};
-  }
-
-  return compactRuntimeConfig({
-    includeComments: true,
-    maxCommentedPosts: maxItemsForLane(lane),
-    maxCommentsPerPost: readNumberParameter(lane, "maxCommentsPerPost"),
-    commentDepth: readNumberParameter(lane, "commentDepth"),
-    commentSort: readStringParameter(lane, "commentSort"),
-  });
-};
-
 const xSearchQueryBudgets = (
   lanes: readonly SourceQueryPlanLane[],
   maxQueries: number,
@@ -476,24 +219,6 @@ const xSearchQueryBudgets = (
     maxItems: budgetsByQuery.get(query) ?? 1,
   }));
 };
-
-const skippedLaneWarnings = (
-  lanes: readonly SourceQueryPlanLane[],
-  executableCount: number,
-): readonly string[] =>
-  executableCount >=
-  lanes.filter((lane) => lane.operation !== "enrichment").length
-    ? []
-    : ["source_query_planner.some_lanes_skipped"];
-
-const fallbackCompilation = (
-  sourceQuery: SourceQuery,
-  warning: string,
-): SourceQueryPlannerRuntimeCompilation => ({
-  sourceQuery,
-  warnings: [warning],
-  applied: false,
-});
 
 const readPlannerAccounts = (
   defaultSourceKey: string,
@@ -576,61 +301,3 @@ const isCommunityListing = (
   value: string | undefined,
 ): value is NonNullable<SourceQueryPlannerCommunity["listings"]>[number] =>
   value === "top" || value === "hot" || value === "new";
-
-const readStringArrayFromValues = (
-  ...values: readonly unknown[]
-): readonly string[] =>
-  compactUnique(
-    values.flatMap(readArray).flatMap((value) => {
-      const text = readString(value);
-
-      return text === undefined ? [] : [text];
-    }),
-  );
-
-const readStringArrayParameter = (
-  lane: SourceQueryPlanLane,
-  key: string,
-): readonly string[] | undefined => {
-  const value = lane.parameters?.[key];
-
-  if (!Array.isArray(value)) {
-    return undefined;
-  }
-
-  const items = value.flatMap((item) => {
-    const text = readString(item);
-
-    return text === undefined ? [] : [text];
-  });
-
-  return items.length === 0 ? undefined : compactUnique(items);
-};
-
-const readStringParameter = (
-  lane: SourceQueryPlanLane,
-  key: string,
-): string | undefined => readString(lane.parameters?.[key]);
-
-const readNumberParameter = (
-  lane: SourceQueryPlanLane,
-  key: string,
-): number | undefined => {
-  const value = lane.parameters?.[key];
-
-  return typeof value === "number" && Number.isFinite(value)
-    ? value
-    : undefined;
-};
-
-const totalMaxItems = (lanes: readonly SourceQueryPlanLane[]): number =>
-  Math.max(
-    1,
-    Math.min(
-      100,
-      lanes.reduce((total, lane) => total + maxItemsForLane(lane), 0),
-    ),
-  );
-
-const maxItemsForLane = (lane: SourceQueryPlanLane): number =>
-  Math.max(1, Math.min(100, lane.maxItems));
