+++ libs/summary/domain/services/reader-summary-topic-label-vocabulary.ts
@@ -1,0 +1,221 @@
+export const weakTopicLabelTokens: ReadonlySet<string> = new Set([
+  "a",
+  "about",
+  "after",
+  "all",
+  "an",
+  "and",
+  "any",
+  "are",
+  "as",
+  "ask",
+  "at",
+  "be",
+  "been",
+  "being",
+  "but",
+  "by",
+  "before",
+  "best",
+  "biggest",
+  "can",
+  "clueless",
+  "could",
+  "council",
+  "day",
+  "days",
+  "did",
+  "didn",
+  "do",
+  "does",
+  "entire",
+  "every",
+  "for",
+  "from",
+  "global",
+  "go",
+  "goto",
+  "has",
+  "have",
+  "having",
+  "here",
+  "hn",
+  "how",
+  "if",
+  "impression",
+  "in",
+  "into",
+  "is",
+  "it",
+  "its",
+  "just",
+  "many",
+  "me",
+  "minute",
+  "minutes",
+  "more",
+  "my",
+  "new",
+  "news",
+  "normal",
+  "not",
+  "one",
+  "of",
+  "often",
+  "on",
+  "or",
+  "over",
+  "people",
+  "post",
+  "posts",
+  "price",
+  "prompt",
+  "prompts",
+  "professionals",
+  "pros",
+  "re",
+  "rely",
+  "relying",
+  "serious",
+  "should",
+  "ship",
+  "shipping",
+  "show",
+  "shown",
+  "source",
+  "story",
+  "that",
+  "the",
+  "these",
+  "this",
+  "those",
+  "thread",
+  "today",
+  "to",
+  "top",
+  "update",
+  "updates",
+  "user",
+  "users",
+  "we",
+  "was",
+  "were",
+  "what",
+  "when",
+  "where",
+  "who",
+  "why",
+  "with",
+  "without",
+  "workshop",
+  "workers",
+  "words",
+  "would",
+  "you",
+  "your",
+]);
+
+export const genericActionTokens: ReadonlySet<string> = new Set([
+  "add",
+  "adds",
+  "added",
+  "adding",
+  "accepted",
+  "according",
+  "announce",
+  "announces",
+  "announced",
+  "behind",
+  "breaking",
+  "build",
+  "built",
+  "building",
+  "bring",
+  "bringing",
+  "brings",
+  "caught",
+  "change",
+  "changes",
+  "come",
+  "comes",
+  "compare",
+  "compares",
+  "consume",
+  "consuming",
+  "created",
+  "creates",
+  "creating",
+  "deep",
+  "ditching",
+  "dropped",
+  "expect",
+  "expected",
+  "every",
+  "forget",
+  "forgot",
+  "gets",
+  "getting",
+  "give",
+  "gave",
+  "gives",
+  "happens",
+  "hiring",
+  "hate",
+  "increase",
+  "decrease",
+  "introduce",
+  "introduced",
+  "introducing",
+  "let",
+  "love",
+  "just",
+  "lead",
+  "leads",
+  "launch",
+  "launched",
+  "launches",
+  "launching",
+  "make",
+  "made",
+  "makes",
+  "many",
+  "migrate",
+  "migrated",
+  "migrates",
+  "migrating",
+  "race",
+  "races",
+  "racing",
+  "report",
+  "reported",
+  "release",
+  "released",
+  "releases",
+  "releasing",
+  "replacing",
+  "relies",
+  "rise",
+  "rises",
+  "rolling",
+  "run",
+  "running",
+  "presents",
+  "say",
+  "said",
+  "says",
+  "showed",
+  "showing",
+  "shows",
+  "ship",
+  "ships",
+  "shipping",
+  "smartest",
+  "test",
+  "tests",
+  "use",
+  "using",
+  "verify",
+  "verifies",
+  "verified",
+  "verifying",
+]);
