+++ ops/release/persistence-readiness-contract.json
@@ -17,19 +17,17 @@
       "moduleFile": "libs/monitoring/interfaces/rest/monitoring-rest.module.ts",
       "context": "monitoring-api",
       "owner": "monitoring-owner",
-    "inMemoryStateAdapters": [
-      "InMemoryTopicRepository",
-      "InMemorySourceBindingRepository",
-      "InMemorySourceCredentialRepository",
-      "InMemorySourceCredentialSecretVault",
-      "InMemoryScanPolicyRepository",
+      "inMemoryStateAdapters": [
+        "InMemoryInterestRepository",
+        "InMemorySourceBindingRepository",
+        "InMemorySourceCredentialRepository",
+        "InMemorySourceCredentialSecretVault",
+        "InMemoryScanPolicyRepository",
         "InMemoryScanJobRepository",
         "InMemoryQueuePublisher",
-        "InMemoryScanQueueAdapter",
-        "InMemoryOutboxAdapter",
         "InMemoryIdempotencyAdapter"
       ],
-      "risk": "topics, source bindings, scan policies, scan jobs, outbox events and idempotency records can opt into Prisma, and scan command publishing can opt into RabbitMQ, but default local mode is still process-local",
+      "risk": "interests, source bindings, credentials, scan policies, scan jobs and idempotency records can opt into durable adapters, but default local mode is still process-local",
       "durableReplacementPlan": "Monitoring repositories, outbox events and idempotency records can opt into Prisma with MONITORING_PERSISTENCE=prisma; scan command publishing can opt into RabbitMQ with MONITORING_SCAN_QUEUE=rabbitmq; event-relay can dispatch Prisma outbox events to RabbitMQ with start:event-relay; remaining blocker is deployment evidence with shared DATABASE_URL/RABBITMQ_URL"
     },
     {
@@ -147,11 +145,13 @@
         "InMemoryScanAttemptRepository",
         "InMemoryScanCursorRepository",
         "InMemorySourceItemRepository",
+        "InMemorySourceCandidateMemoryRepository",
+        "InMemoryConversationUnitRepository",
         "InMemoryScanFailureQueueAdapter",
         "InMemoryQueuePublisher",
         "InMemoryScanCommandQueueReader"
       ],
-      "risk": "worker scan attempts, source items, cursors, leases, failures, projected feed and default scan command reader are process-local",
+      "risk": "worker scan attempts, source items, source-candidate memory, conversation units, cursors, leases, failures, projected feed and default scan command reader are process-local",
       "durableReplacementPlan": "IngestionWorkerModule can opt into Prisma source item, cursor, feed projection, scan failure queue, scan attempt and scan lease persistence with INGESTION_WORKER_PERSISTENCE=prisma; scan command reading can opt into RabbitMQ with INGESTION_SCAN_QUEUE_READER=rabbitmq; platform inbox dedupe can use PrismaInboxStoreAdapter when event consumers are wired through it; check:cross-process-scheduler proves Prisma-backed scheduling into RabbitMQ command publication; remaining blocker is deployed shared-infra evidence"
     },
     {
